requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.5.1
django-coverage-plugin==1.5.0


Sphinx==1.7.4
django-extensions==2.0.7
Werkzeug==0.14.1
django-test-plus==1.0.22
factory-boy==2.11.1


django-debug-toolbar==1.9.1

# improved REPL
ipdb==0.11

pytest-django==3.2.1
pytest-sugar==0.9.1
